finished

+++ src/main.rs
@@ -1,6 +1,6 @@
 #![feature(new_range_api)]
+use rvvm::{Bus, Cpu, Memory};
 use std::io::Read;
-use rvvm::{Cpu, Memory, Bus};
 
 use clap::Parser;
 
@@ -13,7 +13,7 @@ struct Args {
     name: String,
     /// Number of times to greet
     #[arg(short, long, default_value_t = 0)]
-    offset: u64,
+    offset: usize,
     /// Number of times to greet
     #[arg(short, long, action, default_value_t = false)]
     verbose: bool,
@@ -22,9 +22,9 @@ fn main() -> anyhow::Result<()> {
     let args = Args::parse();
     let mut buffer = Vec::new();
     std::fs::File::open(&args.name)
-    .map_err(|err|anyhow::anyhow!("{} {}", err, args.name))?
-    .read_to_end(&mut buffer)?;
-    let mut mem = Memory::new(0u64..=buffer.len() as u64);
+        .map_err(|err| anyhow::anyhow!("{} {}", err, args.name))?
+        .read_to_end(&mut buffer)?;
+    let mut mem = Memory::new(0usize..=buffer.len());
     mem.init_from(&buffer)?;
     let mut c = Cpu::new(mem);
     c.set_pc(args.offset);
